Fix a few warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@25683 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/pdf/pdfobject.cpp
@@ -94,8 +94,8 @@ static CPLString GDALPDFGetPDFString(const char* pszStr)
         else
 #endif
         {
-            osStr += CPLSPrintf("%02X", (pwszDest[i] >> 8) & 0xff);
-            osStr += CPLSPrintf("%02X", (pwszDest[i]) & 0xff);
+            osStr += CPLSPrintf("%02X", (int)(pwszDest[i] >> 8) & 0xff);
+            osStr += CPLSPrintf("%02X", (int)(pwszDest[i]) & 0xff);
         }
     }
     osStr += ">";
